+++ src/main/java/org/kpmp/spatialViewerDataset/SpatialViewerFileDatasetRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -14,5 +15,5 @@ public interface SpatialViewerFileDatasetRepository
     List<SpatialViewerFileDataset> findAll();
 
     @Query (value = "SELECT MAX(release_version) AS max FROM sv_file_v", nativeQuery = true)
-    public Double max();
+    public BigDecimal max();
 }
